+++ NEWS.md
@@ -1,3 +1,4 @@
+- `session_info()` now only lists loaded versions for namespaces which are already loaded.
 
 # 0.0.0.9000